reduce duplication in on-time components and pages
