+++ frontend-two/components/on-time/OnTimeServicesTable.tsx
@@ -1,13 +1,19 @@
 import Link from "next/link";
-import { Duration } from "luxon";
 import { useCallback, useMemo, useState, type ReactNode } from "react";
 import { CsvExportButton } from "@/components/shared/CsvExportButton";
 import { SortedPaginatedTable } from "@/components/table/SortedPaginatedTable";
 import type { SortableTableRow } from "@/components/table/SortableTable";
 import { FrequentServicePerformance } from "@/services/on-time/performance.service";
 import { FrequentIcon } from "../icons/FrequentIcon";
-
-export type ServiceDisplayMode = "percentage" | "count" | "time";
+import {
+  type OnTimeDisplayMode,
+  formatPercentage,
+  formatDuration,
+  formatDirection,
+  formatMetricValue,
+  getMetricSortValue,
+  aggregatePerformanceTotals,
+} from "@/utils/on-time-table-format";
 
 const ALL_COLUMNS = [
   {
@@ -49,183 +55,10 @@ const SERVICE_EXPORT_HEADER_LABELS: Record<string, string> = {
   early: "Early",
 };
 
-function formatDelay(delay: number | null | undefined): string {
-  if (delay == null) return "-";
-  const roundedDelay = Math.round(delay);
-  return (
-    (roundedDelay >= 0 ? "+" : "-") +
-    Duration.fromObject({ seconds: Math.abs(roundedDelay) }).toFormat("mm:ss")
-  );
-}
-
-function formatPercentage(ratio: number | null | undefined): string {
-  if (ratio == null) return "-";
-  return `${(ratio * 100).toFixed(1)}%`;
-}
-
-function formatSeconds(value: number | null | undefined): string {
-  if (value == null) return "-";
-
-  const rounded = Math.round(value);
-  return (
-    (rounded >= 0 ? "+" : "-") +
-    Duration.fromObject({ seconds: Math.abs(rounded) }).toFormat("mm:ss")
-  );
-}
-
-function formatMetricValue(
-  row: FrequentServicePerformance,
-  metric: "onTime" | "late" | "early",
-  displayMode: ServiceDisplayMode,
-): string {
-  const hasActualDepartures = (row.actualDepartures ?? 0) > 0;
-  if (!hasActualDepartures) return "-";
-
-  switch (displayMode) {
-    case "count":
-      return (row[metric] ?? 0).toLocaleString();
-    case "time": {
-      const secondsKey = `${metric}InSeconds` as const;
-      return formatSeconds(row[secondsKey]);
-    }
-    case "percentage":
-    default: {
-      const ratioKey = `${metric}Ratio` as const;
-      return formatPercentage(row[ratioKey]);
-    }
-  }
-}
-
-function getMetricSortValue(
-  row: FrequentServicePerformance,
-  metric: "onTime" | "late" | "early",
-  displayMode: ServiceDisplayMode,
-): string | number {
-  const hasActualDepartures = (row.actualDepartures ?? 0) > 0;
-  if (!hasActualDepartures) return Number.NEGATIVE_INFINITY;
-
-  switch (displayMode) {
-    case "count":
-      return row[metric] ?? Number.NEGATIVE_INFINITY;
-    case "time": {
-      const secondsKey = `${metric}InSeconds` as const;
-      return row[secondsKey] ?? Number.NEGATIVE_INFINITY;
-    }
-    case "percentage":
-    default: {
-      const ratioKey = `${metric}Ratio` as const;
-      return row[ratioKey] ?? Number.NEGATIVE_INFINITY;
-    }
-  }
-}
-
-function formatDirection(direction: string | null | undefined): string {
-  if (!direction) return "-";
-  else if (direction.toLowerCase() === "inbound") return "Inbound";
-  else if (direction.toLowerCase() === "outbound") return "Outbound";
-  return "-";
-}
-
-function calculateTotals(
-  data: FrequentServicePerformance[],
-): FrequentServicePerformance | null {
-  if (data.length === 0) return null;
-
-  let scheduledDepartures = 0;
-  let actualDepartures = 0;
-  let onTime = 0;
-  let late = 0;
-  let early = 0;
-  let total = 0;
-  let onTimeRatioSum = 0;
-  let lateRatioSum = 0;
-  let earlyRatioSum = 0;
-  let countDelayed = 0;
-  let weightedDelayTotal = 0;
-  let hasDelayData = false;
-  let onTimeInSecondsTotal = 0;
-  let onTimeInSecondsCount = 0;
-  let lateInSecondsTotal = 0;
-  let lateInSecondsCount = 0;
-  let earlyInSecondsTotal = 0;
-  let earlyInSecondsCount = 0;
-
-  for (const row of data) {
-    scheduledDepartures += row.scheduledDepartures ?? 0;
-    actualDepartures += row.actualDepartures ?? 0;
-    onTime += row.onTime ?? 0;
-    late += row.late ?? 0;
-    early += row.early ?? 0;
-    total += row.total ?? 0;
-    onTimeRatioSum += row.onTimeRatio ?? 0;
-    lateRatioSum += row.lateRatio ?? 0;
-    earlyRatioSum += row.earlyRatio ?? 0;
-
-    if (row.averageDelay != null || row.countDelayed != null) {
-      hasDelayData = true;
-      countDelayed += row.countDelayed ?? 0;
-      weightedDelayTotal += (row.averageDelay ?? 0) * (row.countDelayed ?? 0);
-    }
-
-    if (row.onTimeInSeconds != null) {
-      onTimeInSecondsTotal += row.onTimeInSeconds;
-      onTimeInSecondsCount += 1;
-    }
-    if (row.lateInSeconds != null) {
-      lateInSecondsTotal += row.lateInSeconds;
-      lateInSecondsCount += 1;
-    }
-    if (row.earlyInSeconds != null) {
-      earlyInSecondsTotal += row.earlyInSeconds;
-      earlyInSecondsCount += 1;
-    }
-  }
-
-  const totalRatio = onTimeRatioSum + lateRatioSum + earlyRatioSum;
-
-  return {
-    ...data[0],
-    lineId: "",
-    lineInfo: {
-      ...data[0].lineInfo,
-      serviceId: "",
-      serviceNumber: "",
-      serviceName: "",
-    },
-    direction: null,
-    scheduledDepartures,
-    actualDepartures,
-    onTime,
-    late,
-    early,
-    total,
-    completedRatio:
-      scheduledDepartures > 0 ? actualDepartures / scheduledDepartures : 0,
-    onTimeRatio: totalRatio > 0 ? onTimeRatioSum / totalRatio : 0,
-    lateRatio: totalRatio > 0 ? lateRatioSum / totalRatio : 0,
-    earlyRatio: totalRatio > 0 ? earlyRatioSum / totalRatio : 0,
-    averageDelay:
-      hasDelayData && countDelayed > 0
-        ? weightedDelayTotal / countDelayed
-        : null,
-    countDelayed: hasDelayData ? countDelayed : null,
-    onTimeInSeconds:
-      onTimeInSecondsCount > 0
-        ? onTimeInSecondsTotal / onTimeInSecondsCount
-        : null,
-    lateInSeconds:
-      lateInSecondsCount > 0 ? lateInSecondsTotal / lateInSecondsCount : null,
-    earlyInSeconds:
-      earlyInSecondsCount > 0
-        ? earlyInSecondsTotal / earlyInSecondsCount
-        : null,
-  };
-}
-
 function getRowValue(
   row: FrequentServicePerformance,
   column: string,
-  displayMode: ServiceDisplayMode,
+  displayMode: OnTimeDisplayMode,
 ): string | number {
   switch (column) {
     case "frequent":
@@ -257,7 +90,7 @@ function getRowValue(
   }
 }
 
-function createRenderRow(nocCode: string, displayMode: ServiceDisplayMode) {
+function createRenderRow(nocCode: string, displayMode: OnTimeDisplayMode) {
   return (row: FrequentServicePerformance): SortableTableRow => {
     const completedRatio =
       row.completedRatio ??
@@ -287,7 +120,7 @@ function createRenderRow(nocCode: string, displayMode: ServiceDisplayMode) {
         displayMode === "percentage"
           ? formatPercentage(completedRatio)
           : (row.actualDepartures ?? 0).toLocaleString(),
-      averageDelay: formatDelay(row.averageDelay),
+      averageDelay: formatDuration(row.averageDelay),
       onTime: formatMetricValue(row, "onTime", displayMode),
       late: formatMetricValue(row, "late", displayMode),
       early: formatMetricValue(row, "early", displayMode),
@@ -298,7 +131,7 @@ function createRenderRow(nocCode: string, displayMode: ServiceDisplayMode) {
 interface OnTimeServicesTableProps {
   data: FrequentServicePerformance[];
   nocCode: string;
-  displayMode: ServiceDisplayMode;
+  displayMode: OnTimeDisplayMode;
   visibleColumns?: string[];
 }
 
@@ -320,12 +153,8 @@ export const OnTimeServicesTable = ({
     [visibleColumns],
   );
   const totalsRow = useMemo<SortableTableRow | null>(() => {
-    const totals = calculateTotals(data);
+    const totals = aggregatePerformanceTotals(data);
     if (!totals) return null;
-
-    const completedRatio =
-      totals.completedRatio ??
-      ((totals.actualDepartures ?? 0) / (totals.scheduledDepartures ?? 0) || 0);
 
     return {
       key: "__totals__",
@@ -333,16 +162,16 @@ export const OnTimeServicesTable = ({
       service: "",
       direction: "-",
       scheduledDepartures: (
-        <strong>{(totals.scheduledDepartures ?? 0).toLocaleString()}</strong>
+        <strong>{totals.scheduledDepartures.toLocaleString()}</strong>
       ),
       recordedDepartures: (
         <strong>
           {displayMode === "percentage"
-            ? formatPercentage(completedRatio)
-            : (totals.actualDepartures ?? 0).toLocaleString()}
+            ? formatPercentage(totals.completedRatio)
+            : totals.actualDepartures.toLocaleString()}
         </strong>
       ),
-      averageDelay: <strong>{formatDelay(totals.averageDelay)}</strong>,
+      averageDelay: <strong>{formatDuration(totals.averageDelay)}</strong>,
       onTime: (
         <strong>{formatMetricValue(totals, "onTime", displayMode)}</strong>
       ),
@@ -379,7 +208,7 @@ export const OnTimeServicesTable = ({
         displayMode === "percentage"
           ? formatPercentage(completedRatio)
           : row.actualDepartures ?? 0,
-      averageDelay: formatDelay(row.averageDelay),
+      averageDelay: formatDuration(row.averageDelay),
       onTime: formatMetricValue(row, "onTime", displayMode),
       late: formatMetricValue(row, "late", displayMode),
       early: formatMetricValue(row, "early", displayMode),
